Refactored and added implicit wrapper
